requirements.txt
python-dotenv
sqlalchemy
asyncpg
alembic
python-telegram-bot==20.5
bcrypt
